app/requirements.txt
streamlit
requests
python-dotenv
langchain
langchain-ollama